moving confinement to its own package
