+++ packages/confinement/.vscode/launch.json
@@ -1,0 +1,98 @@
+{
+  // Use IntelliSense to learn about possible attributes.
+  // Hover to view descriptions of existing attributes.
+  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+  "version": "0.2.0",
+  "inputs": [
+    {
+      "type": "promptString",
+      "id": "HUC",
+      "description": "What HUC?",
+      // "default": "17070202" // North Fork John Day
+      "default": "17060304" // Really small HUC
+      // "default": "17060103" // Asotin
+    }
+  ],
+  "configurations": [
+    {
+      "name": "Write Confinement report",
+      "type": "python",
+      "request": "launch",
+      "module": "confinement.confinement_report",
+      "console": "integratedTerminal",
+      "args": [
+        "{env:DATA_ROOT}/confinement/${input:HUC}/outputs/brat.sqlite",
+        "{env:DATA_ROOT}/confinement/${input:HUC}/project.rs.xml",
+        "{env:DATA_ROOT}/confinement/${input:HUC}/outputs/brat_report.html"
+      ]
+    },
+    {
+      "name": "Version Bump",
+      "type": "python",
+      "module": "rscommons.version",
+      "args": [
+        "${workspaceFolder}/confinement/__version__.py"
+      ],
+      "request": "launch",
+      "console": "integratedTerminal"
+    },
+    {
+      "name": "Confinement",
+      "type": "python",
+      "request": "launch",
+      "module": "confinement.confinement",
+      "console": "integratedTerminal",
+      // Add NATIONAL_PROJECT, INPUTS_FOLDER, DOWNLOAD_FOLDER to your .env file
+      "args": [
+        "${input:HUC}",
+        "{env:DATA_ROOT}/vbet/${input:HUC}/inputs/vbet_inputs.gpkg/flowlines_vaa",
+        "{env:DATA_ROOT}/channel_area/${input:HUC}/outputs/channel_area.gpkg/channel_area",
+        "{env:DATA_ROOT}/vbet/${input:HUC}/outputs/vbet.gpkg/vbet_full",
+        "{env:DATA_ROOT}/confinement/${input:HUC}", // output folder
+        "vbet_level_path",
+        "ValleyBottom",
+        "--buffer", "15.0",
+        "--segmented_network", "{env:DATA_ROOT}/rs_context/${input:HUC}/hydrology/hydro_derivatives.gpkg/network_intersected_300m",
+        "--meta",
+        "Runner=VSCode",
+        "--verbose",
+        //"--debug"
+      ]
+    },
+    {
+      "name": "Calculate Confinement",
+      "type": "python",
+      "request": "launch",
+      "module": "rme.confinement",
+      "console": "integratedTerminal",
+      // Add NATIONAL_PROJECT, INPUTS_FOLDER, DOWNLOAD_FOLDER to your .env file
+      "args": [
+        "${input:HUC}",
+        "{env:DATA_ROOT}/vbet/${input:HUC}/inputs/vbet_inputs.gpkg/Flowlines_VAA",
+        "{env:DATA_ROOT}/channel_area/${input:HUC}/outputs/channel_area.gpkg/channel_area",
+        "{env:DATA_ROOT}/vbet/${input:HUC}/outputs/vbet.gpkg/vbet_full",
+        "{env:DATA_ROOT}/confinement/${input:HUC}", // output folder
+        "vbet_level_path",
+        "ValleyBottom",
+        "--buffer", "15.0",
+        "--segmented_network", "{env:DATA_ROOT}/rs_context/${input:HUC}/hydrology/hydrology.gpkg/network_intersected_300m",
+        "--calculate_existing",
+        "--meta",
+        "Runner=VSCode",
+        "--verbose",
+        //"--debug"
+      ]
+    },
+    {
+      "name": "Augment Confinement XML",
+      "type": "python",
+      "request": "launch",
+      "module": "rme.confinement_rs",
+      "console": "integratedTerminal",
+      "args": [
+        "{env:DATA_ROOT}/confinement/${input:HUC}/project.rs.xml",
+        "{env:DATA_ROOT}/rs_context/${input:HUC}/project.rs.xml,{env:DATA_ROOT}/vbet/${input:HUC}/project.rs.xml"
+      ]
+    }
+  ]
+}
