+++ packages/confinement/confinement/confinement.py
@@ -26,8 +26,8 @@ from rscommons import GeopackageLayer
 from rscommons.vector_ops import collect_feature_class, get_geometry_unary_union, copy_feature_class
 from rscommons.util import safe_makedirs, parse_metadata
 from rme.shapley_ops import line_segments, select_geoms_by_intersection, cut
-from rme.utils.confinement_report import ConfinementReport
-from rme.__version__ import __version__
+from confinement.confinement_report import ConfinementReport
+from confinement.__version__ import __version__
 
 Path = str
