+++ packages/confinement/confinement/docs/.gitignore
@@ -1,0 +1,11 @@
+*.gem
+.bundle
+.sass-cache
+_site
+.jekyll-cache
+Gemfile.lock
+node_modules
+
+.jekyll-metadata
+.DS_Store
+resources/pocket\.md
