+++ packages/confinement/confinement/docs/Gemfile
@@ -1,0 +1,9 @@
+# A note about this GEM file:
+# This gemfile is only to help you run `bundle exec jekyll serve` on your
+# local machine. It is not used in any way by the build system
+# frozen_string_literal: true
+
+source "https://rubygems.org"
+gem "jekyll", "3.9.0"
+gem "jekyll-remote-theme", "0.4.2"
+gem "kramdown-parser-gfm", "1.1.0"
