+++ packages/confinement/confinement/docs/_config.yml
@@ -1,0 +1,79 @@
+# Site title will be used everywhere
+title:  Confinement
+description: > # this means to ignore newlines until "baseurl:"
+  Online help and documentation site for Confinement
+destination: ../../../PUBLIC/confinement
+# Final root url of the site. No ending slash. No subdirectory
+url:
+# If the site is in a subdirectory you should do it like this: "/subdirectory". Otherwise ""
+baseurl: "/confinement"
+# When serving locally what port should we use:
+port: 4001
+
+# Markdown style
+markdown: kramdown
+# Syntax highlighter.
+highlighter: rouge
+
+# Choosing "default" layout means we don't need a "layout:default" in all our 
+# front matter
+defaults:
+  -
+    scope:
+      path: "" # an empty string here means all files in the project
+    values:
+      layout: "default"
+
+# Files/Folders to exclude from publishing
+exclude:
+  - vendor
+  - src
+  - LICENSE
+  - README.md
+  - _config.yml
+  - Gemfile*
+
+remote_theme: Riverscapes/riverscapes-jekyll-theme
+plugins:
+  - jekyll-remote-theme
+
+# These are theme settings that will be used to turn features on and off
+# They are not part of Jekyll, they are entirely custom.
+settings:
+  # Turn the automatically generated page contents on and off
+  autoPageContents:
+    active: true
+    onHome: false
+  # Options for the sidebar menu      
+  sideMenu:
+    homeItem: true
+    startExpanded: false
+  # Arbitrary multi-level topbar menu
+  # IF you don't need this just comment out everything below
+  # ----------------------------------
+  topmenu:
+    - title: "USU Watershed Sciences Department"
+      url: https://qcnr.usu.edu/wats/
+    - title: "LTPBR"
+      url: http://lowtechpbr.restoration.usu.edu/
+    - title: "Riverscapes Consortium"
+      url: https://riverscapes.net
+    - title: "Tools"
+      url: https://riverscapes.net/Tools/
+    - title: "Data"
+      url: https://data.riverscapes.net
+    - title: "Maps"
+      url: https://maps.riverscapes.net
+
+
+# Redirecting from apps can be tricky so we use a key-value system
+# key: must be in the set [A-Za-z0-9_] (no spaces, no dashes)
+# url: must be relative url, no leading slash.
+# 
+# The url you have to call is: http://siteurl.com/baseurl/?APPKEY=KEY1
+# ---------------------------------------------------------------
+#AppRedirects:
+#  - key: KEY1
+#    url: ThingsA
+#  - key: KEY2
+#    url: ThingsB/subpageB.html
