+++ packages/confinement/confinement/docs/_includes/sidebar.html
@@ -1,0 +1,3 @@
+
+<!-- This is where you put your extra sidebar info like a mailchimp form -->
+<a href="/">Back to riverscapes tools</a>
