+++ packages/confinement/confinement/docs/index.md
@@ -1,0 +1,5 @@
+---
+title: Home
+---
+
+Confinement Dolor duis esse velit ullamco ea ut proident incididunt. Cillum velit laboris aute veniam. Ipsum mollit commodo amet consequat esse laboris eiusmod commodo aliqua sint. Ea elit eu commodo pariatur occaecat laborum elit magna enim.
