+++ packages/confinement/confinement/layer_descriptions.json
@@ -1,0 +1,5 @@
+{
+    "JUNCTION_POINTS": ["","",""],
+    "POINT_METRICS": ["","",""],
+    "POINT_MEASUREMENTS": ["","",""]
+}
