+++ packages/rme/.vscode copy/launch.json
@@ -1,0 +1,134 @@
+{
+  // Use IntelliSense to learn about possible attributes.
+  // Hover to view descriptions of existing attributes.
+  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+  "version": "0.2.0",
+  "inputs": [
+    {
+      "type": "promptString",
+      "id": "HUC",
+      "description": "What HUC?",
+      // "default": "17070202" // North Fork John Day
+      "default": "17060304" // Really small HUC
+      // "default": "17060103" // Asotin
+    }
+  ],
+  "configurations": [
+    {
+      "name": "🚀 Run RME",
+      "type": "python",
+      "request": "launch",
+      "module": "rme.metric_engine",
+      "console": "integratedTerminal",
+      // Add NATIONAL_PROJECT, INPUTS_FOLDER, DOWNLOAD_FOLDER to your .env file
+      "args": [
+        "${input:HUC}",
+        "{env:DATA_ROOT}/rs_context/${input:HUC}/hydrology/nhdplushr.gpkg/NHDFlowline",
+        "{env:DATA_ROOT}/rs_context/${input:HUC}/hydrology/nhdplushr.gpkg/NHDPlusFlowlineVAA",
+        "{env:DATA_ROOT}/vbet/${input:HUC}/intermediates/vbet_intermediates.gpkg/vbet_dgos",
+        "{env:DATA_ROOT}/vbet/${input:HUC}/outputs/vbet.gpkg/vbet_igos",
+        "{env:DATA_ROOT}/vbet/${input:HUC}/outputs/vbet.gpkg/vbet_centerlines",
+        "{env:DATA_ROOT}/rs_context/${input:HUC}/topography/dem.tif",
+        "{env:DATA_ROOT}/rs_context/${input:HUC}/climate/precipitation.tif",
+        "{env:DATA_ROOT}/rs_context/${input:HUC}/transportation/roads.shp",
+        "{env:DATA_ROOT}/rs_context/${input:HUC}/transportation/railways.shp",
+        "{env:DATA_ROOT}/rs_context/${input:HUC}/ecoregions/ecoregions.shp",
+        "{env:DATA_ROOT}/rme/${input:HUC}",
+        //"--epsg", "5070",
+        //"--meta",
+        "--verbose"
+      ]
+    },
+    {
+      "name": "Write Confinement report",
+      "type": "python",
+      "request": "launch",
+      "module": "sqlbrat.brat_report",
+      "console": "integratedTerminal",
+      "args": [
+        "{env:DATA_ROOT}/confinement/${input:HUC}/outputs/brat.sqlite",
+        "{env:DATA_ROOT}/confinement/${input:HUC}/project.rs.xml",
+        "{env:DATA_ROOT}/confinement/${input:HUC}/outputs/brat_report.html"
+      ]
+    },
+    {
+      "name": "Version Bump",
+      "type": "python",
+      "module": "rscommons.version",
+      "args": [
+        "${workspaceFolder}/rme/__version__.py"
+      ],
+      "request": "launch",
+      "console": "integratedTerminal"
+    },
+    {
+      "name": "Confinement",
+      "type": "python",
+      "request": "launch",
+      "module": "rme.confinement",
+      "console": "integratedTerminal",
+      // Add NATIONAL_PROJECT, INPUTS_FOLDER, DOWNLOAD_FOLDER to your .env file
+      "args": [
+        "${input:HUC}",
+        "{env:DATA_ROOT}/vbet/${input:HUC}/inputs/vbet_inputs.gpkg/flowlines_vaa",
+        "{env:DATA_ROOT}/channel_area/${input:HUC}/outputs/channel_area.gpkg/channel_area",
+        "{env:DATA_ROOT}/vbet/${input:HUC}/outputs/vbet.gpkg/vbet_full",
+        "{env:DATA_ROOT}/confinement/${input:HUC}", // output folder
+        "vbet_level_path",
+        "ValleyBottom",
+        "--buffer", "15.0",
+        "--segmented_network", "{env:DATA_ROOT}/rs_context/${input:HUC}/hydrology/hydro_derivatives.gpkg/network_intersected_300m",
+        "--meta",
+        "Runner=VSCode",
+        "--verbose",
+        //"--debug"
+      ]
+    },
+    {
+      "name": "Calculate Confinement",
+      "type": "python",
+      "request": "launch",
+      "module": "rme.confinement",
+      "console": "integratedTerminal",
+      // Add NATIONAL_PROJECT, INPUTS_FOLDER, DOWNLOAD_FOLDER to your .env file
+      "args": [
+        "${input:HUC}",
+        "{env:DATA_ROOT}/vbet/${input:HUC}/inputs/vbet_inputs.gpkg/Flowlines_VAA",
+        "{env:DATA_ROOT}/channel_area/${input:HUC}/outputs/channel_area.gpkg/channel_area",
+        "{env:DATA_ROOT}/vbet/${input:HUC}/outputs/vbet.gpkg/vbet_full",
+        "{env:DATA_ROOT}/confinement/${input:HUC}", // output folder
+        "vbet_level_path",
+        "ValleyBottom",
+        "--buffer", "15.0",
+        "--segmented_network", "{env:DATA_ROOT}/rs_context/${input:HUC}/hydrology/hydrology.gpkg/network_intersected_300m",
+        "--calculate_existing",
+        "--meta",
+        "Runner=VSCode",
+        "--verbose",
+        //"--debug"
+      ]
+    },
+    {
+      "name": "Augment Confinement XML",
+      "type": "python",
+      "request": "launch",
+      "module": "rme.confinement_rs",
+      "console": "integratedTerminal",
+      "args": [
+        "{env:DATA_ROOT}/confinement/${input:HUC}/project.rs.xml",
+        "{env:DATA_ROOT}/rs_context/${input:HUC}/project.rs.xml,{env:DATA_ROOT}/vbet/${input:HUC}/project.rs.xml"
+      ]
+    },
+    {
+      "name": "Augment RME XML",
+      "type": "python",
+      "request": "launch",
+      "module": "rme.rme_rs",
+      "console": "integratedTerminal",
+      "args": [
+        "{env:DATA_ROOT}/rme/${input:HUC}/project.rs.xml",
+        "{env:DATA_ROOT}/rs_context/${input:HUC}/project.rs.xml,{env:DATA_ROOT}/vbet/${input:HUC}/project.rs.xml"
+      ]
+    }
+  ]
+}
