+++ packages/rme/.vscode copy/settings.json
@@ -1,0 +1,27 @@
+{
+  "[python]": {
+    "editor.tabSize": 4,
+    "editor.formatOnSave": true,
+  },
+  "python.linting.enabled": true,
+  "python.linting.lintOnSave": true,
+  "python.linting.pycodestyleEnabled": true,
+  "python.linting.pylintEnabled": true,
+  "python.linting.pycodestyleArgs": [
+    "--ignore=E501"
+  ],
+  "python.formatting.autopep8Args": [
+    "--ignore=E501"
+  ],
+  "python.terminal.activateEnvironment": true,
+  "python.autoComplete.showAdvancedMembers": false,
+  "files.exclude": {
+    "**/*.egg-info": true
+  },
+  "files.watcherExclude": {
+    "**/*.egg-info/**": true
+  },
+  "search.exclude": {
+    "**/*.egg-info": true
+  },
+}
